test/runner: Add python script to automate test execution
Implement python script which automates the execution of all test cases
tithin the main test folder. For each test, it performs two separate
runs: one utilizing the cluster with 1 core, the second with 8. The
results of each run are saved in a .csv file in 'test/runner/result/'
directory. Upon completion of all tests, the script processes these data
files to generate the "benchmnark.md" file, presenting the performance
statistics in a comparative table format.